+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Display quotes table page with seller quotes if the quote was splitted
+
+## [1.6.6] - 2025-01-06
+### Fixed
+- Fix Request Quote message in Quotes Details
+
 ## [1.6.5] - 2024-09-26
 ### Fixed